Refactor `CsvFileConnector` and `CsvDataSource`.
