+++ src/main/java/edu/ie3/datamodel/io/csv/CsvFileDefinition.java
@@ -5,6 +5,12 @@
 */
 package edu.ie3.datamodel.io.csv;
 
+import edu.ie3.datamodel.exceptions.FileException;
+import edu.ie3.datamodel.io.naming.FileNamingStrategy;
+import edu.ie3.datamodel.models.Entity;
+import edu.ie3.datamodel.models.timeseries.TimeSeries;
+import edu.ie3.datamodel.models.timeseries.TimeSeriesEntry;
+import edu.ie3.datamodel.models.value.Value;
 import edu.ie3.datamodel.utils.FileUtils;
 import java.nio.file.Path;
 import java.util.Arrays;
@@ -21,6 +27,67 @@ public record CsvFileDefinition(Path filePath, String[] headLineElements, String
   public CsvFileDefinition(
       String fileName, Path directoryPath, String[] headLineElements, String csvSep) {
     this(FileUtils.ofCsv(fileName, directoryPath), headLineElements, csvSep);
+  }
+
+  /**
+   * Builds a new file definition consisting of file name and headline elements
+   *
+   * @param clz Class that is meant to be serialized into this file
+   * @param headLineElements Array of headline elements
+   * @param csvSep Separator for csv columns
+   * @param fileNamingStrategy that should be used
+   * @throws FileException If the definition cannot be determined
+   */
+  public CsvFileDefinition(
+      Class<? extends Entity> clz,
+      String[] headLineElements,
+      String csvSep,
+      FileNamingStrategy fileNamingStrategy)
+      throws FileException {
+    this(
+        FileUtils.ofCsv(
+            fileNamingStrategy
+                .getEntityName(clz)
+                .orElseThrow(
+                    () ->
+                        new FileException(
+                            "Cannot determine the file name for class '"
+                                + clz.getSimpleName()
+                                + "'.")),
+            fileNamingStrategy.getDirectoryPath(clz).orElse(Path.of(""))),
+        headLineElements,
+        csvSep);
+  }
+
+  /**
+   * Builds a new file definition consisting of file name and headline elements
+   *
+   * @param timeSeries Time series to derive naming information from
+   * @param headLineElements Array of headline elements
+   * @param csvSep Separator for csv columns
+   * @param fileNamingStrategy that should be used
+   * @throws FileException If the definition cannot be determined
+   */
+  public <T extends TimeSeries<E, V>, E extends TimeSeriesEntry<V>, V extends Value>
+      CsvFileDefinition(
+          T timeSeries,
+          String[] headLineElements,
+          String csvSep,
+          FileNamingStrategy fileNamingStrategy)
+          throws FileException {
+    this(
+        FileUtils.ofCsv(
+            fileNamingStrategy
+                .getEntityName(timeSeries)
+                .orElseThrow(
+                    () ->
+                        new FileException(
+                            "Cannot determine the file name for time series '"
+                                + timeSeries
+                                + "'.")),
+            fileNamingStrategy.getDirectoryPath(timeSeries).orElse(Path.of(""))),
+        headLineElements,
+        csvSep);
   }
 
   /**
