+++ src/main/java/edu/ie3/datamodel/io/source/csv/CsvWeatherSource.java
@@ -139,10 +139,7 @@ public class CsvWeatherSource extends WeatherSource {
       throws SourceException {
     /* Get only weather time series meta information */
     Collection<CsvIndividualTimeSeriesMetaInformation> weatherCsvMetaInformation =
-        dataSource
-            .connector
-            .getCsvIndividualTimeSeriesMetaInformation(ColumnScheme.WEATHER)
-            .values();
+        dataSource.getCsvIndividualTimeSeriesMetaInformation(ColumnScheme.WEATHER).values();
     return readWeatherTimeSeries(Set.copyOf(weatherCsvMetaInformation), dataSource.connector);
   }
