+++ src/test/groovy/edu/ie3/datamodel/io/csv/CsvFileDefinitionTest.groovy
@@ -5,11 +5,24 @@
  */
 package edu.ie3.datamodel.io.csv
 
+import edu.ie3.datamodel.exceptions.FileException
+import edu.ie3.datamodel.io.naming.DefaultDirectoryHierarchy
+import edu.ie3.datamodel.io.naming.EntityPersistenceNamingStrategy
+import edu.ie3.datamodel.io.naming.FileNamingStrategy
+import edu.ie3.datamodel.models.StandardUnits
+import edu.ie3.datamodel.models.input.NodeInput
+import edu.ie3.datamodel.models.timeseries.individual.IndividualTimeSeries
+import edu.ie3.datamodel.models.timeseries.individual.TimeBasedValue
+import edu.ie3.datamodel.models.timeseries.repetitive.RepetitiveTimeSeries
+import edu.ie3.datamodel.models.value.EnergyPriceValue
 import org.apache.commons.io.FilenameUtils
 import spock.lang.Shared
 import spock.lang.Specification
+import tech.units.indriya.quantity.Quantities
 
+import java.nio.file.Files
 import java.nio.file.Path
+import java.time.ZonedDateTime
 
 class CsvFileDefinitionTest extends Specification {
   @Shared
@@ -72,6 +85,103 @@ class CsvFileDefinitionTest extends Specification {
       assert it.directoryPath == this.directory
       assert it.headLineElements() == this.headLineElements
       assert it.csvSep() == this.csvSep
+    }
+  }
+
+  def "The csv file definition can be build correctly from class upon request"() {
+    given:
+    def fileNamingStrategy = new FileNamingStrategy()
+    def expected = new CsvFileDefinition("node_input.csv", Path.of(""), ["a", "b", "c"] as String[], ",")
+
+    when:
+    def actual = new CsvFileDefinition(NodeInput, ["a", "b", "c"] as String[], ",", fileNamingStrategy)
+
+    then:
+    actual.with {
+      assert it.filePath == expected.filePath
+      assert it.headLineElements() == expected.headLineElements()
+      assert it.csvSep() == expected.csvSep()
+    }
+  }
+
+  def "The csv file connector can be build correctly from class upon request, utilizing directory hierarchy"() {
+    given:
+    def tmpDirectory = Files.createTempDirectory("psdm_csv_file_")
+    def fileNamingStrategy = new FileNamingStrategy(new EntityPersistenceNamingStrategy(), new DefaultDirectoryHierarchy(tmpDirectory, "test"))
+    def expected = new CsvFileDefinition("node_input.csv", Path.of("test", "input", "grid"), ["a", "b", "c"] as String[], ",")
+
+    when:
+    def actual = new CsvFileDefinition(NodeInput, ["a", "b", "c"] as String[], ",", fileNamingStrategy)
+
+    then:
+    actual.with {
+      assert it.filePath == expected.filePath
+      assert it.headLineElements() == expected.headLineElements()
+      assert it.csvSep() == expected.csvSep()
+    }
+  }
+
+  def "The csv file definition throws FileException if it cannot be build from time series"() {
+    given:
+    def fileNamingStrategy = new FileNamingStrategy()
+
+    and: "credible input"
+    def timeSeries = Mock(RepetitiveTimeSeries)
+
+    when:
+    new CsvFileDefinition(timeSeries, ["a", "b", "c"] as String[], ",", fileNamingStrategy)
+
+    then:
+    def ex = thrown(FileException)
+    ex.message == "Cannot determine the file name for time series 'Mock for type 'RepetitiveTimeSeries' named 'timeSeries''."
+  }
+
+  def "The csv file definition can be build correctly from time series upon request"() {
+    given:
+    def fileNamingStrategy = new FileNamingStrategy()
+    def expected = new CsvFileDefinition("its_c_0c03ce9f-ab0e-4715-bc13-f9d903f26dbf.csv", Path.of(""), ["a", "b", "c"] as String[], ",")
+
+    and: "credible input"
+    def entries = [
+      new TimeBasedValue(ZonedDateTime.now(), new EnergyPriceValue(Quantities.getQuantity(50d, StandardUnits.ENERGY_PRICE)))
+    ] as SortedSet
+    def timeSeries = Mock(IndividualTimeSeries)
+    timeSeries.uuid >> UUID.fromString("0c03ce9f-ab0e-4715-bc13-f9d903f26dbf")
+    timeSeries.entries >> entries
+
+    when:
+    def actual = new CsvFileDefinition(timeSeries, ["a", "b", "c"] as String[], ",", fileNamingStrategy)
+
+    then:
+    actual.with {
+      assert it.filePath == expected.filePath
+      assert it.headLineElements() == expected.headLineElements()
+      assert it.csvSep() == expected.csvSep()
+    }
+  }
+
+  def "The csv file definition can be build correctly from time series upon request, utilizing directory hierarchy"() {
+    given:
+    def tmpDirectory = Files.createTempDirectory("psdm_csv_file_")
+    def fileNamingStrategy = new FileNamingStrategy(new EntityPersistenceNamingStrategy(), new DefaultDirectoryHierarchy(tmpDirectory, "test"))
+    def expected = new CsvFileDefinition("its_c_0c03ce9f-ab0e-4715-bc13-f9d903f26dbf.csv", Path.of("test", "input", "participants", "time_series"), ["a", "b", "c"] as String[], ",")
+
+    and: "credible input"
+    def entries = [
+      new TimeBasedValue(ZonedDateTime.now(), new EnergyPriceValue(Quantities.getQuantity(50d, StandardUnits.ENERGY_PRICE)))
+    ] as SortedSet
+    def timeSeries = Mock(IndividualTimeSeries)
+    timeSeries.uuid >> UUID.fromString("0c03ce9f-ab0e-4715-bc13-f9d903f26dbf")
+    timeSeries.entries >> entries
+
+    when:
+    def actual = new CsvFileDefinition(timeSeries, ["a", "b", "c"] as String[], ",", fileNamingStrategy)
+
+    then:
+    actual.with {
+      assert it.filePath == expected.filePath
+      assert it.headLineElements() == expected.headLineElements()
+      assert it.csvSep() == expected.csvSep()
     }
   }
 }
